🎨  refactor(actions): move challenge select ui
